Refactor FormErrors component according to BEM rules
